+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changes to this project are documented in this file. More detail and links can b
 
 ## 1.16.0
 ### Added
+- AUTOTOOL-1652: Add support for routing as-path
 - AUTOTOOL-1374: ([GitHub Issue 112](https://github.com/F5Networks/f5-declarative-onboarding/issues/112)): Add support for MirrorIp class
 ### Fixed
 - Target VLAN errors from the inspect endpoint